feat(ci.j, trusted.ci and cert.ci) migrate to separated tf modules for agents management in Azure (#516)
This PR is a (mandatory) preparatory step for
https://github.com/jenkins-infra/helpdesk/issues/3818 .

It split the controller/azurevm-agents/aci-agent scopes in 3 terraform
modules. The goal is to allow instantiating the (*)agents module for the
new sponsorship subscrption without repeating code.


To avoid any breakage on the principal branch (which uses the latest
reference of the `main` branch on
https://github.com/jenkins-infra/shared-tools/), I've created 3 brand
new modules in
https://github.com/jenkins-infra/shared-tools/commit/c7ec5b0b350abfb195f191056ad6877073aa1e7c
. It should make the PR here autonomous to merge (but the aformentionned
commit is also to be reviewed and we can update modules).


💡 A few notes on the introduces changes:

- The ci.jenkins.io's Network Security Group rule
`allow_outbound_ssh_from_ci_controller_to_s390x` is removed. Its
integrated into the new controller module as one of the agent IP
prefixes passed as argument.
- Same for the trusted.ci.jenkins.io's
`allow_outbound_ssh_from_controller_to_permanent_agent`
- The 3 Network Security Group rules
`allow_inbound_ssh_from_controller_to_ephemeral_agents` (1 for each
controller) are changed from a single `source_address_prefix` to
`source_address_prefixes` collection
- For ci.jenkins.io, it also adds the Public IP of the controller VM in
this collection (along to the private VM IP) to cover cases where the
requests are routed through the Internet instead of the internal network
peerings

---------

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>