hotfix(ci.jenkins.io) use correct provider for the `maven-cache` Azure File Share (#990)
Fixup of #985 #987 and #988

Ref. https://github.com/jenkins-infra/helpdesk/issues/4619


The failure to locate the azure file comes from a missing `provider` on
the Azure File Share resource in the TF code.
The share has been created but cannot be read as it searches in a
distinct subscription (Azure CDF instead of Azure Sponsored).