fixup(privatek8s-sponsorship) use the same private CIDR for pods as in other private clusters (#1005)
Ref. https://github.com/jenkins-infra/azure/pull/997/files#r2037035929.

As per
https://learn.microsoft.com/en-us/azure/aks/concepts-network-azure-cni-overlay,
the pod IPs are NAT-ed. So we can reuse the same CIDR on all clusters.

The constraint is to be sure that, at maximum node capacity, the CIDR
can have a `/24` carved out for each node.
Since this sizing fits ci-jenkins-io-agents-1, we should not worry here.

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>