fix: ensure all azurerm_storage_account are network-allowed for all agents (using local) (#1072)
While pairing on https://github.com/jenkins-infra/helpdesk/issues/4689,
we realized that the subnets and allowed IPs of our controllers are
spread which can lead to build errors when manipulating network
resources.

This PR improves the subnet specification per controller component to
improve the specification and centralize the definition of which subnet
is used by which controller and component.

Note: it is a fix as it adds the subnet used by the [new
`infracijenkinsioagents2` AKS
cluster](https://github.com/jenkins-infra/azure/pull/1057) to all azure
storage account's allowed networks

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>