feat(trusted.ci) create resources for ephemeral agents to migrate from sponsored subscription (#1107)
This PR creates the resources to sustain ephemeral agent Azure VM for
trusted.ci in the Azure CDF subscription (instead of the Azure Sponsored
subscription).

We do not delete resources until we have moved and verified workloads.
Exception of 2 resources recreated due to a renaming in the "Private
Endpoint" module use to allow the trusted permanent agent to reach the
mirrorbits API with the `mirrorbits` CLI.

---------

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>