fix(dockerhub-mirror) use existing private DNS zone for trusted.ci  (#1172)
Fixup of
https://github.com/jenkins-infra/azure/pull/1171#issuecomment-3199486226

Note: terratest are annoying here as they show the weakness of terraform
plan on an empty platform when a lot of "dynamic" loops are used. As
such, we do not "add" some resources when running in terratest to avoid
errors.

This PR also adds NSG rules for trusted <-> ACR (missing in #1171)

---------

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>