fix(trusted.ci) use the Azure automatic private DNS zones instead of our custom one (#1173)
Related to https://github.com/jenkins-infra/helpdesk/issues/4769

In order to allow trusted.ci to utilized the private ACR, it must use
the "automatic" private DNS zone.
Alas we can only associate one private DNS zone per virtual network.

This PR deletes the current zone and update records to be set in the new
one, OR in the public zone for trusted.ci.

---------

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>