Revert "feat(publick8s) create a new cluster with modern (2025) setup" (#1188)
Reverts jenkins-infra/azure#1187 which failed to apply: NAT gateway do
not work with Ipv6.

The error messages pointed to
https://learn.microsoft.com/en-us/azure/nat-gateway/faq#can-an-ipv6-public-ip-address-be-used-with-a-nat-gateway
which suggest to use a combination of egress methods on AKS when using
dual stack.

We'll re-create the cluster creation with this technique on a subsequent
PR. As such, let's rollback