feat(publick8s) create a new cluster with modern (2025) setup (#1190)
Ref.
https://github.com/jenkins-infra/helpdesk/issues/4617#issuecomment-3295780800

New attempt as #1187 failed and was roll backed in #1188 with the
following changes:

- ~Use AzureLinux3 as per Azure's depreciation email~ (Failed to specify
`AzureLinux3` -
https://github.com/hashicorp/terraform-provider-azurerm/issues/30575 so
we stick with the default)
- Set up the cluster's default egress method to a load balancer with
same settings as before: IPv6 requires the LB as [NAT gateway do not
support
IPv6](https://learn.microsoft.com/en-us/azure/nat-gateway/faq#can-an-ipv6-public-ip-address-be-used-with-a-nat-gateway).
- Ipv4 also goes through the LB as it is the *main* egress method. IF
the LB fails, then NAT gateway will be used (and IPv6 will fail)

---------

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>