feat(weekly.ci,publick8s) set up a new RG and empty disk with proper naming convention (#1203)
Related to https://github.com/jenkins-infra/helpdesk/issues/4617

This PR creates a new RG and a new disk for weekly.ci.jenkins.io:

- The new disk is empty, and will be populated twice: once with the
initial data from old cluster `publick8s-endless-ghoul` and an second
time at migration time.
- This is a process to ensure we can fully start the service in the new
`publick8s` cluster without putting the old service (which is production
until we migrate). We'll use this process (if it is working as expected)
for LDAP (which also uses a disk)
- Only the new cluster will use this disk, instead of the old one: it
will recreate permissions, PV and PVC of course.

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>