fix(release.ci.jenkins.io/privatek8s) ensure the release.ci data azure file storage has a correct volumeHandle (#1221)
Following
https://github.com/jenkins-infra/helpdesk/issues/4813#issuecomment-3339518665
and a pre-requisite to
https://github.com/jenkins-infra/helpdesk/issues/4817, this PR fixes the
configuration of the Azure File CSI Persistent Volume (PV) used by
release.ci.jenkins.io in the cluster `privatek8s` for data storage from
agents.

It introduces the following changes:

- chore: move the PV/PVC definition in the `privatek8s.tf` file, closer
to the cluster definition. It reverts a change I made a few months
earlier as it proves easier to manage.
- feat: add a claimRef on the changed PV: it ensures that only 1 PVC can
claim it so we control this kind of association.
- cleanup: remove an unused storage class (as everything is statically
provisioned)
- fix: set up the `spec.csi.volumeHandle` of the PV to use the exact
convention expected by the Azure File CSI driver 1.32.x+


Note: before merging this PR, the PV and PVC must be removed manually to
ensure that the re-creation works as expected.
Then, we should use release.ci.jenkins.io controller to trigger the
"Infra Agents Health" job to verify that the PV/PVC can be used by
agents as usual (to avoid failure on the next weekly release).

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>