feat(agent.trusted.ci.jenkins.io) Add a new SSD data disk (to replace HDD on long term) (#1228)
Related to
https://github.com/jenkins-infra/helpdesk/issues/4828#issuecomment-3381045275

This PR adds a new SSD disk with the characteristics described in the
issue.
Once attached, we'll manually mount it and start filling it with the
same content as the current HDD data disk. The goal is to replace the
HDD once we're ready (see issue for following the data copy steps)

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>