cleanup(archives.jenkins.io) migrate DNS record to jenkins-infra/azure-net (#1243)
While working on
https://github.com/jenkins-infra/helpdesk/issues/4695#issuecomment-3479913280,
we relaized that it was easier to migrate the `archives.jenkins.io` and
`archives.jenkins-ci.org` records to jenkins-infra/azure-net.

The first step is this PR which removes the code and tells Terraform to
NOT delete the records (but only stop tracking them).

Once applied:

- https://github.com/jenkins-infra/azure-net/pull/461 will be mergeable
(to import the 2 records and eventually changes them)
- A subsequent PR to cleanup the `removed {}` block after the records
are fully migrated

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>